chore(ruff): clear all 659 lint errors + add ruff config (#61)
Resolve all 659 preexisting ruff errors. Repo now passes 'ruff check .'.

- Add [tool.ruff] config with documented rules and per-file-ignores.
- Auto-fix I001, RUF100, UP017/022/035, SIM105/117, FLY002, FURB.
- chmod +x on 91 CLI scripts (EXE001).
- Fix B904, RUF012, DTZ901, SIM117 manually.
- Guard import-order dependency in test_diagnostic_sanitization.py.
- Drop _vault_graduation.py +x (library, no shebang).

Rules ignored with justification: E501, BLE001, DTZ005, B008, SIM102/103/115,
E702, F403, F401 (re-exports without __all__).

Validation: gates passed, 653 tests pass, ruff clean.

+++ scripts/evals/hook_latency_benchmark.py
@@ -10,21 +10,20 @@ import tempfile
 import time
 from pathlib import Path
 
-
 REPO_ROOT = Path(__file__).resolve().parents[2]
 HOOKS_ROOT = REPO_ROOT / ".codex" / "hooks"
 if str(HOOKS_ROOT) not in sys.path:
     sys.path.insert(0, str(HOOKS_ROOT))
 
-from shared.active_context import active_context_from_payload  # noqa: E402
-from shared.autoresearch_observer import observe_post_tool_payload  # noqa: E402
+from shared.active_context import active_context_from_payload
+from shared.autoresearch_observer import observe_post_tool_payload
 
 
 def percentile(values: list[float], pct: float) -> float:
     if not values:
         return 0.0
     ordered = sorted(values)
-    index = min(len(ordered) - 1, int(round((pct / 100.0) * (len(ordered) - 1))))
+    index = min(len(ordered) - 1, round((pct / 100.0) * (len(ordered) - 1)))
     return ordered[index]
